requirements.txt
azure-cognitiveservices-vision-face==0.4.0
click==6.7
Flask==1.0.2
Flask-Admin==1.5.3
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
MarkupSafe==1.0
opencv-python==4.1.0.25
Pillow==5.1.0
requests==2.21.0
SQLAlchemy==1.3.2
urllib3==1.24.1
Werkzeug==0.14.1
WTForms==2.2.1
